Clean up MetaMetaStrategy based on linter advice (#2)
* Add initial tests to MetaMetaStrategy

* CLean up MetaMetaStrategy based on linter advice

This addresses some of the linter concerns for the MetaMetaStrategy
class. It breaks some of the more complex methods down, to hush the
branch condition linter. It also does some re-indenting and removes some
print statements.

* Further housekeeping changes to MetaMetaStrategy

* MostFrequentBot added

This bot plays the counter move to that of the opponents most frequent
move.

* Fix MetaMetaStrategy initializer after bad rebase